Address new comments
