+++ parquet/src/main/java/org/apache/iceberg/data/parquet/InternalReader.java
@@ -20,13 +20,10 @@ package org.apache.iceberg.data.parquet;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.StructLike;
-import org.apache.iceberg.data.GenericRecord;
 import org.apache.iceberg.parquet.ParquetValueReader;
 import org.apache.iceberg.parquet.ParquetValueReaders;
-import org.apache.iceberg.parquet.ParquetValueReaders.StructReader;
 import org.apache.iceberg.types.Types.StructType;
 import org.apache.parquet.column.ColumnDescriptor;
 import org.apache.parquet.schema.LogicalTypeAnnotation;
@@ -39,12 +36,12 @@ public class InternalReader extends BaseParquetReaders<StructLike> {
 
   private InternalReader() {}
 
-  public static ParquetValueReader<StructLike> buildReader(
+  public static ParquetValueReader<StructLike> create(
       Schema expectedSchema, MessageType fileSchema) {
     return INSTANCE.createReader(expectedSchema, fileSchema);
   }
 
-  public static ParquetValueReader<StructLike> buildReader(
+  public static ParquetValueReader<StructLike> create(
       Schema expectedSchema, MessageType fileSchema, Map<Integer, ?> idToConstant) {
     return INSTANCE.createReader(expectedSchema, fileSchema, idToConstant);
   }
@@ -52,79 +49,38 @@ public class InternalReader extends BaseParquetReaders<StructLike> {
   @Override
   protected ParquetValueReader<StructLike> createStructReader(
       List<Type> types, List<ParquetValueReader<?>> fieldReaders, StructType structType) {
-    return new StructLikeReader(types, fieldReaders, structType);
+    return new ParquetValueReaders.RecordReader<>(types, fieldReaders, structType);
   }
 
   @Override
-  protected ParquetValueReaders.PrimitiveReader<?> fixedReader(ColumnDescriptor desc) {
+  protected ParquetValueReader<?> fixedReader(ColumnDescriptor desc) {
     return new ParquetValueReaders.BytesReader(desc);
   }
 
   @Override
-  protected ParquetValueReaders.PrimitiveReader<?> int96Reader(ColumnDescriptor desc) {
-    return new ParquetValueReaders.TimestampInt96Reader(desc);
+  protected ParquetValueReader<?> dateReader(ColumnDescriptor desc) {
+    return new ParquetValueReaders.UnboxedReader<>(desc);
   }
 
   @Override
-  protected Optional<ParquetValueReader<?>> dateReader(ColumnDescriptor desc) {
-    return Optional.of(new ParquetValueReaders.UnboxedReader<>(desc));
-  }
-
-  @Override
-  protected Optional<ParquetValueReader<?>> timeReader(
+  protected ParquetValueReader<?> timeReader(
       ColumnDescriptor desc, LogicalTypeAnnotation.TimeUnit unit) {
     if (unit == LogicalTypeAnnotation.TimeUnit.MILLIS) {
-      return Optional.of(new ParquetValueReaders.TimestampMillisReader(desc));
+      return ParquetValueReaders.timestampMillisReader(desc);
     }
 
-    return Optional.of(new ParquetValueReaders.UnboxedReader<>(desc));
+    return new ParquetValueReaders.UnboxedReader<>(desc);
   }
 
   @Override
-  protected Optional<ParquetValueReader<?>> timestampReader(
+  protected ParquetValueReader<?> timestampReader(
       ColumnDescriptor desc, LogicalTypeAnnotation.TimeUnit unit, boolean isAdjustedToUTC) {
     if (unit == LogicalTypeAnnotation.TimeUnit.MILLIS) {
-      return Optional.of(new ParquetValueReaders.TimestampMillisReader(desc));
+      return ParquetValueReaders.timestampMillisReader(desc);
+    } else if (unit == LogicalTypeAnnotation.TimeUnit.NANOS) {
+      return ParquetValueReaders.timestampInt96Reader(desc);
     }
 
-    return Optional.of(new ParquetValueReaders.UnboxedReader<>(desc));
-  }
-
-  @Override
-  protected Optional<ParquetValueReader<?>> uuidReader(ColumnDescriptor desc) {
-    return Optional.of(new ParquetValueReaders.UUIDReader(desc));
-  }
-
-  private static class StructLikeReader extends StructReader<StructLike, StructLike> {
-    private final GenericRecord template;
-
-    StructLikeReader(List<Type> types, List<ParquetValueReader<?>> readers, StructType struct) {
-      super(types, readers);
-      this.template = struct != null ? GenericRecord.create(struct) : null;
-    }
-
-    @Override
-    protected StructLike newStructData(StructLike reuse) {
-      if (reuse != null) {
-        return reuse;
-      } else {
-        return template.copy();
-      }
-    }
-
-    @Override
-    protected Object getField(StructLike intermediate, int pos) {
-      return intermediate.get(pos, Object.class);
-    }
-
-    @Override
-    protected StructLike buildStruct(StructLike struct) {
-      return struct;
-    }
-
-    @Override
-    protected void set(StructLike struct, int pos, Object value) {
-      struct.set(pos, value);
-    }
+    return new ParquetValueReaders.UnboxedReader<>(desc);
   }
 }
